DMT: Fix for cursor locking not always locking
git-svn-id: svn://svn.code.sf.net/p/dualmonitortool/code/trunk@457 fb2ba5e6-08d8-4d4c-bdde-3d14dca355d2

+++ DMT/DMT/Library/PInvoke/NativeMethods.cs
@@ -206,7 +206,8 @@ namespace DMT.Library.PInvoke
 			public uint mouseData;
 			public uint flags;
 			public uint time;
-			public uint dwExtraInfo;
+			//public uint dwExtraInfo;
+			public IntPtr dwExtraInfo;
 		}
 
 		[StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
